Move the turning point finding into a tok_geo_utils.c. Also plan to move other chunks of codes with if statements about which type of geo (sol, core, private plasma) into this file. Tested for core and outboard SOL. Will add tests for other regions (update old tests to match with newer tokamak eometry code).
