+++ zero/gkyl_tok_geo_priv.h
@@ -1,7 +1,34 @@
 #include <gkyl_tok_geo.h>
 
+// Function context to pass to root finder
+struct arc_length_ctx {
+  const gkyl_tok_geo *geo;
+  double *arc_memo;
+  double *arc_memo_left;
+  double *arc_memo_right;
+  double psi, rclose, zmin, arcL;
+  double rleft, rright, zmax;
+  double arcL_right; // this is for when we need to switch sides
+  double arcL_left; // this is for when we need to switch sides
+  double arcL_tot; // total arc length
+  double phi_right; // this is for when we need to switch sides
+  double phi_left; // this is for when we need to switch sides
+  double phi_bot; // For new way of trying to do core
+  bool right;
+  double zmaxis;
+  enum gkyl_tok_geo_type ftype; // type of geometry
+};
+
+
+// Context to pass to endpoint finder
+struct plate_ctx{
+  const struct gkyl_tok_geo* geo;
+  double psi_curr;
+  bool lower;
+};
+
 // some helper functions
-double
+static double
 choose_closest(double ref, double* R, double* out, int nr)
 {
   //return fabs(R[0]-ref) < fabs(R[1]-ref) ? out[0] : out[1];
@@ -321,3 +348,8 @@ integrate_phi_along_psi_contour_memo(const gkyl_tok_geo *geo, double psi,
   ((gkyl_tok_geo *)geo)->stat.nquad_cont_calls += ctx.ncall;
   return res;
 }
+
+
+double phi_func(double alpha_curr, double Z, void *ctx);
+double plate_psi_func(double s, void *ctx);
+void find_endpoints(struct gkyl_tok_geo_geo_inp* inp, struct gkyl_tok_geo *geo, struct arc_length_ctx* arc_ctx, struct plate_ctx* pctx, double psi_curr, double alpha_curr, double* zmin, double* zmax, double* arc_memo, double* arc_memo_left, double* arc_memo_right);
