+++ zero/tok_geo.c
@@ -2,7 +2,6 @@
 #include <gkyl_array.h>
 #include <gkyl_array_rio.h>
 #include <gkyl_basis.h>
-#include <gkyl_tok_geo.h>
 #include <gkyl_math.h>
 #include <gkyl_range.h>
 #include <gkyl_rect_grid.h>
@@ -16,30 +15,8 @@
 
 
 
-// Function context to pass to root finder
-struct arc_length_ctx {
-  const gkyl_tok_geo *geo;
-  double *arc_memo;
-  double *arc_memo_left;
-  double *arc_memo_right;
-  double psi, rclose, zmin, arcL;
-  double rleft, rright, zmax;
-  double arcL_right; // this is for when we need to switch sides
-  double arcL_left; // this is for when we need to switch sides
-  double phi_right; // this is for when we need to switch sides
-  double phi_left; // this is for when we need to switch sides
-  double phi_bot; // For new way of trying to do core
-  bool right;
-  double zmaxis;
-  enum gkyl_tok_geo_type ftype; // type of geometry
-};
-struct plate_ctx{
-  const struct gkyl_tok_geo* geo;
-  double psi_curr;
-  bool lower;
-};
 
-static inline double
+double
 plate_psi_func(double s, void *ctx){
   // uses a pointer to the plate function to get R(s), Z(s)
   // Then calculates psi(R, Z)
@@ -129,7 +106,7 @@ arc_length_func(double Z, void *ctx)
 }
 
 // Function to calculate phi given alpha
-static inline double
+double
 phi_func(double alpha_curr, double Z, void *ctx)
 {
   struct arc_length_ctx *actx = ctx;
@@ -397,229 +374,13 @@ void gkyl_tok_geo_calc(struct gk_geometry* up, struct gkyl_range *nrange, double
           double psi_curr = phi_lo + ip*dpsi + modifiers[ip_delta]*delta_psi;
           double zmin = inp->zmin, zmax = inp->zmax;
 
-          double arcL, darcL, arcL_curr, arcL_lo;
+          double darcL, arcL_curr, arcL_lo;
           double arcL_l, arcL_r;
           double phi_r, phi_l;
-          if(inp->ftype == GKYL_CORE){
-            //Find the turning points
-            double zlo, zup, zlo_last;
-            zlo = 0.01;
-            zup=zmax;
-            zlo_last = zlo;
-            double R[4], dR[4];
-            while(true){
-              int nlo = R_psiZ(geo, psi_curr, zlo, 4, R, dR);
-              if(nlo==2){
-                if(fabs(zlo-zup)<1e-12){
-                  zmax = zlo;
-                  break;
-                }
-                zlo_last = zlo;
-                zlo = (zlo+zup)/2;
-              }
-              if(nlo==0){
-                zup = zlo;
-                zlo = zlo_last;
-              }
-            }
-            for(int i =0; i<4; i++){
-              R[i] = 0.0;
-              dR[i] = 0.0;
-            }
-            int nup = 0;
-            double zup_last;
-            zup = -0.01;
-            zlo=zmin;
-            zup_last = zup;
-            while(true){
-              int nup = R_psiZ(geo, psi_curr, zup, 4, R, dR);
-              if(nup==2){
-                if(fabs(zlo-zup)<1e-12){
-                  zmin = zup;
-                  break;
-                }
-                zup_last = zup;
-                zup = (zlo+zup)/2;
-              }
-              if(nup==0){
-                zlo = zup;
-                zup = zup_last;
-              }
-            }
-            // Done finding turning points
-            arcL_r = integrate_psi_contour_memo(geo, psi_curr, zmin, zmax, rright,
-              true, true, arc_memo_right);
-            arc_ctx.arcL_right = arcL_r;
-            arc_ctx.right = false;
-            arcL_l = integrate_psi_contour_memo(geo, psi_curr, zmin, zmax, rleft,
-              true, true, arc_memo_left);
-            arcL = arcL_l + arcL_r;
-            darcL = arcL/(up->basis.poly_order*inp->cgrid.cells[TH_IDX]) * (inp->cgrid.upper[TH_IDX] - inp->cgrid.lower[TH_IDX])/2/M_PI;
 
-            arc_ctx.right = true;
-            arc_ctx.phi_right = 0.0;
-            arc_ctx.rclose = rright;
-            arc_ctx.psi = psi_curr;
-            phi_r = phi_func(alpha_curr, zmax, &arc_ctx);
-            arc_ctx.phi_right = phi_r - alpha_curr; // otherwise alpha will get added on twice
-          }
-          else if(inp->ftype == GKYL_PF_LO){
-            //Find the  upper turning point
-            double zlo, zup, zlo_last;
-            zlo = fmax(inp->zmin_left, inp->zmin_right);
-            zup=zmax;
-            zlo_last = zlo;
-            double R[4], dR[4];
-            while(true){
-              int nlo = R_psiZ(geo, psi_curr, zlo, 4, R, dR);
-              if(nlo>=2){
-                if(fabs(zlo-zup)<1e-12){
-                  zmax = zlo;
-                  break;
-                }
-                zlo_last = zlo;
-                zlo = (zlo+zup)/2;
-              }
-              if(nlo<2){
-                zup = zlo;
-                zlo = zlo_last;
-              }
-            }
-            // Done finding turning point
-            arcL_r = integrate_psi_contour_memo(geo, psi_curr, inp->zmin_right, zmax, rright,
-              true, true, arc_memo);
-            arc_ctx.arcL_right = arcL_r;
-            arc_ctx.right = false;
-            arcL_l = integrate_psi_contour_memo(geo, psi_curr, inp->zmin_left, zmax, rleft,
-              true, true, arc_memo);
-            arcL = arcL_l + arcL_r;
-            darcL = arcL/(up->basis.poly_order*inp->cgrid.cells[TH_IDX]) * (inp->cgrid.upper[TH_IDX] - inp->cgrid.lower[TH_IDX])/2/M_PI;
-
-            arc_ctx.right = true;
-            arc_ctx.phi_right = 0.0;
-            arc_ctx.rclose = rright;
-            arc_ctx.psi = psi_curr;
-            arc_ctx.zmin = inp->zmin_right;
-            phi_r = phi_func(alpha_curr, zmax, &arc_ctx);
-            arc_ctx.phi_right = phi_r - alpha_curr; // otherwise alpha will get added on twice
-          }
-          else if(inp->ftype == GKYL_PF_UP){
-            //Find the lower turning point
-            double zlo, zup, zlo_last;
-            double zup_last;
-            zup = fmin(inp->zmax_left, inp->zmax_right);
-            zlo=zmin;
-            zup_last = zup;
-            double R[4], dR[4];
-            while(true){
-              int nup = R_psiZ(geo, psi_curr, zup, 4, R, dR);
-              if(nup>=2){
-                if(fabs(zlo-zup)<1e-12){
-                  zmin = zup;
-                  break;
-                }
-                zup_last = zup;
-                zup = (zlo+zup)/2;
-              }
-              if(nup<2){
-                zlo = zup;
-                zup = zup_last;
-              }
-            }
-            // Done finding turning point
-            arcL_r = integrate_psi_contour_memo(geo, psi_curr, zmin, inp->zmax_right, rright,
-              true, true, arc_memo);
-            arc_ctx.right = false;
-            arcL_l = integrate_psi_contour_memo(geo, psi_curr, zmin, inp->zmax_left, rleft,
-              true, true, arc_memo);
-            arc_ctx.arcL_left= arcL_l;
-            arcL = arcL_l + arcL_r;
-            darcL = arcL/(up->basis.poly_order*inp->cgrid.cells[TH_IDX]) * (inp->cgrid.upper[TH_IDX] - inp->cgrid.lower[TH_IDX])/2/M_PI;
-
-            arc_ctx.right = false;
-            arc_ctx.phi_left = 0.0;
-            arc_ctx.rclose = rleft;
-            arc_ctx.psi = psi_curr;
-            arc_ctx.zmax = inp->zmax_left;
-            phi_l = phi_func(alpha_curr, zmin, &arc_ctx);
-            arc_ctx.phi_left = phi_l - alpha_curr; // otherwise alpha will get added on twice
-          }
-          else if(inp->ftype==GKYL_SOL_DN_OUT){
-            if (geo->plate_spec){ // if we dont have a fixed zmin and zmax
-              double rzplate[2];
-              pctx.psi_curr = psi_curr;
-              pctx.lower=false;
-              double a = 0;
-              double b = 1;
-              double fa = plate_psi_func(a, &pctx);
-              double fb = plate_psi_func(b, &pctx);
-              struct gkyl_qr_res res = gkyl_ridders(plate_psi_func, &pctx,
-                a, b, fa, fb, geo->root_param.max_iter, 1e-10);
-              double smax = res.res;
-              geo->plate_func_upper(smax, rzplate);
-              zmax = rzplate[1];
-
-              pctx.lower=true;
-              a = 0;
-              b = 1;
-              fa = plate_psi_func(a, &pctx);
-              fb = plate_psi_func(b, &pctx);
-              res = gkyl_ridders(plate_psi_func, &pctx,
-                a, b, fa, fb, geo->root_param.max_iter, 1e-10);
-              double smin = res.res;
-              geo->plate_func_lower(smin, rzplate);
-              zmin = rzplate[1];
-            }
-
-            arc_ctx.phi_right = 0.0;
-            arcL = integrate_psi_contour_memo(geo, psi_curr, zmin, zmax, rclose, true, true, arc_memo);
-            darcL = arcL/(up->basis.poly_order*inp->cgrid.cells[TH_IDX]) * (inp->cgrid.upper[TH_IDX] - inp->cgrid.lower[TH_IDX])/2/M_PI;
-          }
-          else if(inp->ftype==GKYL_SOL_DN_IN){
-            arc_ctx.phi_right = 0.0;
-            arcL = integrate_psi_contour_memo(geo, psi_curr, zmin, zmax, rclose, true, true, arc_memo);
-            darcL = arcL/(up->basis.poly_order*inp->cgrid.cells[TH_IDX]) * (inp->cgrid.upper[TH_IDX] - inp->cgrid.lower[TH_IDX])/2/M_PI;
-          }
-          else if(inp->ftype == GKYL_SOL_SN_LO){
-            //Find the  upper turning point
-            double zlo, zup, zlo_last;
-            zlo = fmax(inp->zmin_left, inp->zmin_right);
-            zup=zmax;
-            zlo_last = zlo;
-            double R[4], dR[4];
-            while(true){
-              int nlo = R_psiZ(geo, psi_curr, zlo, 4, R, dR);
-              if(nlo>=2){
-                if(fabs(zlo-zup)<1e-12){
-                  zmax = zlo;
-                  break;
-                }
-                zlo_last = zlo;
-                zlo = (zlo+zup)/2;
-              }
-              if(nlo<2){
-                zup = zlo;
-                zlo = zlo_last;
-              }
-            }
-            // Done finding turning point
-            arcL_r = integrate_psi_contour_memo(geo, psi_curr, inp->zmin_right, zmax, rright,
-              true, true, arc_memo);
-            arc_ctx.arcL_right = arcL_r;
-            arc_ctx.right = false;
-            arcL_l = integrate_psi_contour_memo(geo, psi_curr, inp->zmin_left, zmax, rleft,
-              true, true, arc_memo);
-            arcL = arcL_l + arcL_r;
-            darcL = arcL/(up->basis.poly_order*inp->cgrid.cells[TH_IDX]) * (inp->cgrid.upper[TH_IDX] - inp->cgrid.lower[TH_IDX])/2/M_PI;
-
-            arc_ctx.right = true;
-            arc_ctx.phi_right = 0.0;
-            arc_ctx.rclose = rright;
-            arc_ctx.psi = psi_curr;
-            arc_ctx.zmin = inp->zmin_right;
-            phi_r = phi_func(alpha_curr, zmax, &arc_ctx);
-            arc_ctx.phi_right = phi_r - alpha_curr; // otherwise alpha will get added on twice
-          }
+          find_endpoints(inp, geo, &arc_ctx, &pctx, psi_curr, alpha_curr, &zmin, &zmax, arc_memo, arc_memo_left, arc_memo_right);
+          double arcL = arc_ctx.arcL_tot;
+          darcL = arcL/(up->basis.poly_order*inp->cgrid.cells[TH_IDX]) * (inp->cgrid.upper[TH_IDX] - inp->cgrid.lower[TH_IDX])/2/M_PI;
 
           // at the beginning of each theta loop we need to reset things
           cidx[PH_IDX] = ip;
@@ -647,7 +408,7 @@ void gkyl_tok_geo_calc(struct gk_geometry* up, struct gkyl_range *nrange, double
               arcL_curr = arcL_lo + it*darcL + modifiers[it_delta]*delta_theta*(arcL/2/M_PI);
               double theta_curr = arcL_curr*(2*M_PI/arcL) - M_PI ; 
               if(inp->ftype==GKYL_CORE){
-                if(arcL_curr <= arcL_r){
+                if(arcL_curr <= arc_ctx.arcL_right){
                   rclose = rright;
                   arc_ctx.right = true;
                   ridders_min = -arcL_curr;
@@ -665,7 +426,7 @@ void gkyl_tok_geo_calc(struct gk_geometry* up, struct gkyl_range *nrange, double
                 }
               }
               if(inp->ftype==GKYL_PF_LO){
-                if(arcL_curr <= arcL_r){
+                if(arcL_curr <= arc_ctx.arcL_right){
                   rclose = rright;
                   arc_ctx.right = true;
                   ridders_min = -arcL_curr;
@@ -683,7 +444,7 @@ void gkyl_tok_geo_calc(struct gk_geometry* up, struct gkyl_range *nrange, double
                 }
               }
               if(inp->ftype==GKYL_PF_UP){
-                if(arcL_curr > arcL_l){
+                if(arcL_curr > arc_ctx.arcL_left){
                   rclose = rright;
                   arc_ctx.right = true;
                   ridders_min = arc_ctx.arcL_left - arcL_curr;
@@ -715,7 +476,7 @@ void gkyl_tok_geo_calc(struct gk_geometry* up, struct gkyl_range *nrange, double
                 arc_ctx.zmax = zmax;
               }
               if(arc_ctx.ftype==GKYL_SOL_SN_LO){
-                if(arcL_curr <= arcL_r){
+                if(arcL_curr <= arc_ctx.arcL_right){
                   rclose = rright;
                   arc_ctx.right = true;
                   ridders_min = -arcL_curr;
